pyproject.toml
[tool.poetry]
name = "bbs-server"
version = "0.1.0"
description = ""
authors = ["kaiiiz <at881005@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
boto3 = "^1.13.22"
SQLAlchemy = "^1.3.17"
python-dotenv = "^0.13.0"
mysqlclient = "^1.4.6"
kafka-python = "^2.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
